refactor: split gemini session parsing and tighten context seams
